Redo the sleep interval calculation
To avoid accumulating drift, always calculate the sleep interval based
on our difference to the next send time. We calculate the next send time
by start_time + sent_packets * send_interval. Conveniently, this means
we can switch to module arithmetic for calculating the sequence number,
hence there's no longer any need to check for overflow when increasing
the sequence number.

While here, refer to the initial sequence number by name, so that it's
clear what we're testing for in assertions. It would be nice to be able
to automatically test wraparound scenarios, but this will have to wait
for some better testing infrastructure.