Print packet loss information even if we don't have an RTT estimate
This can happen in situations with heavy packet loss, where it'll take
some time to receive the first response. In that case, we can still
print out passive updates, though the only thing we know for sure is
that the packet loss is 100%; the other fields have to remain blank.