Print out the address we're trying to ping
Some admins use the traditional ping as a utility for name resolution
too, as it happens to print out the target IP address when it starts up.

Let's also print out the selected IP address, as well as the set of
addresses returned by the DNS server (as a gentle reminder that RR might
be at play).

While here, improve the error messages on name resolution failure.